requirements.txt
fastapi==0.110.0
supabase-py==2.4.0
openai==1.14.2
python-dotenv==1.0.1
uvicorn==0.27.1 